+++ bpm-engine-impl/src/main/java/io/takari/bpm/reducers/ExpressionsReducer.java
@@ -54,19 +54,20 @@ public class ExpressionsReducer extends BpmnErrorHandlingReducer {
 
         final EvalExpressionAction a = (EvalExpressionAction) action;
 
-        Variables vars = VariablesHelper.applyInVariables(contextFactory, state.getVariables(), a.getIn(), a.isCopyAllVariables());
-        final ExecutionContextImpl ctx = contextFactory.create(vars, a.getDefinitionId(), a.getElementId());
-
-        boolean storeResult = cfg.isStoreExpressionEvalResultsInContext();
-        Callable<Command> fn = new DelegateFn(javaDelegateHandler, ctx, a.getType(), a.getExpression(), a.getDefaultCommand(), storeResult);
-
-        List<Timeout<Command>> timeouts = a.getTimeouts();
-        if (timeouts != null && !timeouts.isEmpty()) {
-            // a timeout handling decorator
-            fn = new TimeoutCallable<>(executor, timeouts, fn);
-        }
-
+        ExecutionContextImpl ctx = null;
         try {
+            Variables vars = VariablesHelper.applyInVariables(contextFactory, state.getVariables(), a.getIn(), a.isCopyAllVariables());
+            ctx = contextFactory.create(vars, a.getDefinitionId(), a.getElementId());
+
+            boolean storeResult = cfg.isStoreExpressionEvalResultsInContext();
+            Callable<Command> fn = new DelegateFn(javaDelegateHandler, ctx, a.getType(), a.getExpression(), a.getDefaultCommand(), storeResult);
+
+            List<Timeout<Command>> timeouts = a.getTimeouts();
+            if (timeouts != null && !timeouts.isEmpty()) {
+                // a timeout handling decorator
+                fn = new TimeoutCallable<>(executor, timeouts, fn);
+            }
+
             Command result = fn.call();
 
             CommandStack stack = state.getStack();